Add generic bloom filter index creation support in ingestion spec schema definition to support both real time and batch segments
Summary:
Add generic bloom filter index creation support in ingestion spec schema definition to support both real time and batch segments

Added a flag `createBloomFilterIndex` which defaults to false and can be optionally set to true for any String dimension in the ingestion schema to create bloom filter indexes for a dimension which can be used later by broker hosts to prune segments. The segment pruning logic in broker process will now look for both filters on dimensions having bloom filter indexes created in addition to filters on current partition dimensions that are used by some shard specs like HashBasedShardSpec, SingleDimensionShardSpec and BloomFilterNamedShardSpec. The bloom filter index can be enabled regardless of what shard spec is in use and regardless of whether a segment is created by batch or real time ingestion.

"dimensionsSpec": {
    "dimensions": [
            {"name": "partner_id", "type":"string", "createBloomFilterIndex": true},
            "eventtype",
            "app",
            {"name": "root_pin_id", "type":"string", "createBloomFilterIndex": true},
            "pin_id",
            "contenttype",
            "pinformat"
      ]
 }

Test Plan: Unit test and integration test

Reviewers: O1139 Druid, jgu, itallam

Reviewed By: O1139 Druid, jgu, itallam

Subscribers: jenkins, mleonard, shawncao, #realtime-analytics

Differential Revision: https://phabricator.pinadmin.com/D747062

commit: f1f73d115a0581893dbf4634dbfd84ae905c3c9b

+++ server/src/main/java/org/apache/druid/segment/realtime/appenderator/AppenderatorImpl.java
@@ -40,6 +40,7 @@ import org.apache.commons.lang.mutable.MutableLong;
 import org.apache.druid.client.cache.Cache;
 import org.apache.druid.data.input.Committer;
 import org.apache.druid.data.input.InputRow;
+import org.apache.druid.data.input.impl.DimensionsSpec;
 import org.apache.druid.java.util.common.DateTimes;
 import org.apache.druid.java.util.common.FileUtils;
 import org.apache.druid.java.util.common.IAE;
@@ -62,6 +63,7 @@ import org.apache.druid.segment.QueryableIndex;
 import org.apache.druid.segment.QueryableIndexSegment;
 import org.apache.druid.segment.ReferenceCountingSegment;
 import org.apache.druid.segment.Segment;
+import org.apache.druid.segment.column.ColumnHolder;
 import org.apache.druid.segment.incremental.IncrementalIndexAddResult;
 import org.apache.druid.segment.incremental.IndexSizeExceededException;
 import org.apache.druid.segment.incremental.ParseExceptionHandler;
@@ -847,8 +849,6 @@ public class AppenderatorImpl implements Appenderator
 
     // Use a descriptor file to indicate that pushing has completed.
     final File persistDir = computePersistDir(identifier);
-    final File mergedTarget = new File(persistDir, "merged");
-    final File mergedSupplimentalIndexTarget = new File(persistDir, "merged_supplimental_index");
     final File descriptorFile = computeDescriptorFile(identifier);
 
     // Sanity checks
@@ -881,29 +881,42 @@ public class AppenderatorImpl implements Appenderator
         }
       }
 
-      removeDirectory(mergedTarget);
-      removeDirectory(mergedSupplimentalIndexTarget);
-
-      if (mergedTarget.exists()) {
-        throw new ISE("Merged target[%s] exists after removing?!", mergedTarget);
+      final File mergedIndexTarget = new File(persistDir, "merged");
+      removeDirectory(mergedIndexTarget);
+      if (mergedIndexTarget.exists()) {
+        throw new ISE("Merged target[%s] exists after removing?!", mergedIndexTarget);
       }
 
-      if (mergedSupplimentalIndexTarget.exists()) {
-        throw new ISE("Merged supplimental index target[%s] exists after removing?!",
-                      mergedSupplimentalIndexTarget);
-      }
+      final File mergedIndexFile;
+      final File mergedSupplimentalIndexFile;
 
-      final File mergedFile;
       final long mergeFinishTime;
       final long startTime = System.nanoTime();
       List<QueryableIndex> indexes = new ArrayList<>();
       Closer closer = Closer.create();
       try {
-        for (FireHydrant fireHydrant : sink) {
+        DimensionsSpec dimensionsSpec = schema.getParser() == null ? null : schema.getParser()
+                .getParseSpec()
+                .getDimensionsSpec();
+        boolean hasSupplimentalIndex = IndexMerger.hasSupplimentalIndex(dimensionsSpec);
+        final File mergedSupplimentalIndexTarget = hasSupplimentalIndex ?
+                new File(persistDir, "merged_supplimental_index") : null;
+        if (hasSupplimentalIndex) {
+          removeDirectory(mergedSupplimentalIndexTarget);
+          if (mergedSupplimentalIndexTarget.exists()) {
+            throw new ISE(
+                    "Merged supplimental index target[%s] exists after removing?!",
+                    mergedSupplimentalIndexTarget
+            );
+          }
+        }
 
+        List<String> dimensionNamesHasBloomFilterIndexes = IndexMerger.getDimensionNamesHasBloomFilterIndexes(
+                dimensionsSpec);
+
+        for (FireHydrant fireHydrant : sink) {
           // if batch, swap/persist did not memory map the incremental index, we need it mapped now:
           if (!isOpenSegments()) {
-
             // sanity
             Pair<File, SegmentId> persistedMetadata = persistedHydrantMetadata.get(fireHydrant);
             if (persistedMetadata == null) {
@@ -931,23 +944,32 @@ public class AppenderatorImpl implements Appenderator
           Pair<ReferenceCountingSegment, Closeable> segmentAndCloseable = fireHydrant.getAndIncrementSegment();
           final QueryableIndex queryableIndex = segmentAndCloseable.lhs.asQueryableIndex();
           log.debug("Segment[%s] adding hydrant[%s]", identifier, fireHydrant);
+          IndexMerger.setHasBloomFilterIndexesInColumnCapabilities(
+              dimensionNamesHasBloomFilterIndexes,
+              dimension -> {
+              ColumnHolder columnHolder = queryableIndex.getColumnHolder(dimension);
+              return columnHolder == null ? null : columnHolder.getCapabilities();
+            }
+          );
           indexes.add(queryableIndex);
           closer.register(segmentAndCloseable.rhs);
         }
 
-        mergedFile = indexMerger.mergeQueryableIndex(
+        Pair<File, File> p = indexMerger.mergeQueryableIndex(
             indexes,
             schema.getGranularitySpec().isRollup(),
             schema.getAggregators(),
             schema.getDimensionsSpec(),
-            mergedTarget,
+            mergedIndexTarget,
+            mergedSupplimentalIndexTarget,
             tuningConfig.getIndexSpec(),
             tuningConfig.getIndexSpecForIntermediatePersists(),
             new BaseProgressIndicator(),
             tuningConfig.getSegmentWriteOutMediumFactory(),
             tuningConfig.getMaxColumnsToMerge()
         );
-
+        mergedIndexFile = p.lhs;
+        mergedSupplimentalIndexFile = p.rhs;
         mergeFinishTime = System.nanoTime();
 
         log.debug("Segment[%s] built in %,dms.", identifier, (mergeFinishTime - startTime) / 1000000);
@@ -969,7 +991,8 @@ public class AppenderatorImpl implements Appenderator
           // Kafka indexing task, pushers must use unique file paths in deep storage in order to maintain exactly-once
           // semantics.
           () -> dataSegmentPusher.push(
-              mergedFile,
+              mergedIndexFile,
+              mergedSupplimentalIndexFile,
               segmentToPush,
               useUniquePath
           ),
